API

Modules

LegendDSP.LegendDSP

Types and constants

LegendDSP.DSPConfig

Functions and macros

LegendDSP.create_dsp_config

LegendDSP.dsp_decay_times

LegendDSP.dsp_sg_optimization

LegendDSP.dsp_trap_ft_optimization

LegendDSP.dsp_trap_rt_optimization

LegendDSP.extrema3points

LegendDSP.get_wvf_maximum

LegendDSP.saturation

LegendDSP.tailstats

Documentation

LegendDSP.LegendDSP — Module

LegendDSP

Template for Julia packages.

LegendDSP.DSPConfig — Type

DSPConfig{T <: Real}

Configuration parameters for DSP algorithms.

Fields

enc_pickoff::Quantity{<:T}: pick-off time for ENC noise calculations

bl_mean::NTuple{2, Quantity{<:T}}: fit window for basline extraction

pz_fit::NTuple{2, Quantity{<:T}}: fit window for decay time extraction

t0_threshold::T: ADC threshold for t0 determination

e_grid_rt_trap::StepRangeLen{Quantity{<:T}}: rise time grid scan range for trapezoidal filter

e_grid_ft_trap::StepRangeLen{Quantity{<:T}}: flat-top time grid scan range for trapezoidal filter

e_grid_rt_zac::StepRangeLen{Quantity{<:T}}: rise time grid scan range for ZAC filter

e_grid_ft_zac::StepRangeLen{Quantity{<:T}}: flat-top time grid scan range for ZAC filter

e_grid_rt_cusp::StepRangeLen{Quantity{<:T}}: rise time grid scan range for CUSP filter

e_grid_ft_cusp::StepRangeLen{Quantity{<:T}}: flat-top time grid scan range for CUSP filter

e_grid_rt_sg::StepRangeLen{Quantity{<:T}}: window length grid scan range for SG filter

Examples

using LegendDSP
using Unitful

dsp_config = DSPConfig{Float64}(32.0u"µs", 
(0.0u"µs", 39.0u"µs"), 
(80.0u"µs", 110.0u"µs"), 
5.0, 
7.0u"µs":0.5u"µs":12.0u"µs", 1.0u"µs":0.2u"µs":4.0u"µs",
7.0u"µs":0.5u"µs":12.0u"µs", 1.0u"µs":0.2u"µs":4.0u"µs", 
7.0u"µs":0.5u"µs":12.0u"µs", 1.0u"µs":0.2u"µs":4.0u"µs")

LegendDSP.create_dsp_config — Function

create_dsp_config(dsp_metadata::PropDicts.PropDict)

Create a DSPConfig from a PropDict of DSP metadata.

Arguments

dsp_metadata::PropDicts.PropDict: DSP metadata

Returns

dsp_config::DSPConfig: DSP configuration

LegendDSP.dsp_decay_times — Method

dsp_decay_times(wvfs::AbstractSamples, start::Real, stop::Real)

Get statistics on the logarhithmic of the tail of the wvfs in the interval (start,stop).

LegendDSP.dsp_sg_optimization — Method

dsp_sg_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}, pars_filter::PropDict) where T<:Real

Optimize the Savitzky-Golay filter parameters for a given waveform set.

Returns: - aoe: Array of efficiency values for the given Savitzky-Golay filter parameters - e: Array of energy values for the given Savitzky-Golay filter parameters - blmean: Baseline mean value - blslope: Baseline slope value

LegendDSP.dsp_trap_ft_optimization — Method

dsp_trap_ft_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}, rt::Quantity{T}) where T<:Real

Get energy grid values for given trap grid rise times while varying the flat-top time.

Returns: - e_grid: Array energy values for the given trap rise time grid at a given flat-top time grid.

LegendDSP.dsp_trap_rt_optimization — Method

dsp_trap_rt_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}) where T<:Real

Get ENC noise grid values for given trap grid rise times.

Returns: - enc_trap_grid: Array ENC noise values for the given trap rise time grid

LegendDSP.extrema3points — Method

extrema3points(y1::T,y2::T,y3::T)::T where {T<:AbstractFloat}

Calculate the extrema of a parabola defined by three points.

LegendDSP.get_wvf_maximum — Function

get_wvf_maximum(signal::AbstractSamples, start::Real, stop::Real)
get_wvf_maximum(signal::RDWaveform, start::RealQuantity, stop::RealQuantity)

Get the maximum of a signal in the interval (start,stop) by quadaratic interpolation.

LegendDSP.saturation — Function

tailstats(signal::AbstractSamples, start::Real, stop::Real)
tailstats(signal::RDWaveform, start::RealQuantity, stop::RealQuantity)

Get statistics on the logarhithmic of the tail of a signal in the interval (start,stop).

LegendDSP.tailstats — Function

tailstats(signal::AbstractSamples, start::Real, stop::Real)
tailstats(signal::RDWaveform, start::RealQuantity, stop::RealQuantity)

Get statistics on the logarhithmic of the tail of a signal in the interval (start,stop).