API

Modules

LegendDSP.LegendDSP

Types and constants

LegendDSP.DSPConfig

LegendDSP.IntersectMaximum

Functions and macros

LegendDSP.create_dsp_config

LegendDSP.dsp_cusp_ft_optimization

LegendDSP.dsp_cusp_rt_optimization

LegendDSP.dsp_decay_times

LegendDSP.dsp_icpc

LegendDSP.dsp_puls

LegendDSP.dsp_sg_optimization

LegendDSP.dsp_sipm

LegendDSP.dsp_trap_ft_optimization

LegendDSP.dsp_trap_rt_optimization

LegendDSP.dsp_zac_ft_optimization

LegendDSP.dsp_zac_rt_optimization

LegendDSP.extrema3points

LegendDSP.get_t0

LegendDSP.get_t50

LegendDSP.get_t80

LegendDSP.get_wvf_maximum

LegendDSP.saturation

LegendDSP.tailstats

Documentation

LegendDSP.LegendDSP — Module

LegendDSP

Template for Julia packages.

LegendDSP.DSPConfig — Type

DSPConfig{T <: Real}

Configuration parameters for DSP algorithms.

Fields

enc_pickoff::Quantity{<:T}: pick-off time for ENC noise calculations

bl_mean::NTuple{2, Quantity{<:T}}: fit window for basline extraction

pz_fit::NTuple{2, Quantity{<:T}}: fit window for decay time extraction

t0_threshold::T: ADC threshold for t0 determination

e_grid_rt_trap::StepRangeLen{Quantity{<:T}}: rise time grid scan range for trapezoidal filter

e_grid_ft_trap::StepRangeLen{Quantity{<:T}}: flat-top time grid scan range for trapezoidal filter

e_grid_rt_zac::StepRangeLen{Quantity{<:T}}: rise time grid scan range for ZAC filter

e_grid_ft_zac::StepRangeLen{Quantity{<:T}}: flat-top time grid scan range for ZAC filter

e_grid_rt_cusp::StepRangeLen{Quantity{<:T}}: rise time grid scan range for CUSP filter

e_grid_ft_cusp::StepRangeLen{Quantity{<:T}}: flat-top time grid scan range for CUSP filter

e_grid_rt_sg::StepRangeLen{Quantity{<:T}}: window length grid scan range for SG filter

Examples

using LegendDSP
using Unitful

dsp_config = DSPConfig{Float64}(32.0u"µs", 
(0.0u"µs", 39.0u"µs"), 
(80.0u"µs", 110.0u"µs"), 
5.0, 
7.0u"µs":0.5u"µs":12.0u"µs", 1.0u"µs":0.2u"µs":4.0u"µs",
7.0u"µs":0.5u"µs":12.0u"µs", 1.0u"µs":0.2u"µs":4.0u"µs", 
7.0u"µs":0.5u"µs":12.0u"µs", 1.0u"µs":0.2u"µs":4.0u"µs")

LegendDSP.IntersectMaximum — Type

struct IntersectMaximum <: Function

Finds the intersects of a Y with a threshold and picking the maximum in a given time window. Constructors:

IntersectMaximum(; fields...)

Fields:

mintot::Union{Unitful.Quantity{<:var"#s10"}, var"#s10"} where var"#s10"<:Real: minimum time-over-threshold

maxtot::Union{Unitful.Quantity{<:var"#s10"}, var"#s10"} where var"#s10"<:Real: maximum time-over-threshold for max to appear

LegendDSP.create_dsp_config — Function

create_dsp_config(dsp_metadata::PropDicts.PropDict)

Create a DSPConfig from a PropDict of DSP metadata.

Arguments

dsp_metadata::PropDicts.PropDict: DSP metadata

Returns

dsp_config::DSPConfig: DSP configuration

LegendDSP.dsp_cusp_ft_optimization — Method

dsp_cusp_ft_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}, rt::Quantity{T}) where T<:Real

Get energy grid values for given CUSP grid rise times while varying the flat-top time.

Returns

- `e_grid`: Array energy values for the given CUSP rise time grid at a given flat-top time grid.

LegendDSP.dsp_cusp_rt_optimization — Method

dsp_cusp_rt_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T},; ft::Quantity{T}=4.0u"µs") where T<:Real

Get ENC noise grid values for given CUSP grid rise times.

Returns

- `enc_cusp_grid`: Array ENC noise values for the given CUSP rise time grid

LegendDSP.dsp_decay_times — Method

dsp_decay_times(wvfs::AbstractSamples, start::Real, stop::Real)

Get statistics on the logarhithmic of the tail of the wvfs in the interval (start,stop).

Returns

τ: decay time in µs

LegendDSP.dsp_icpc — Method

dsp_icpc(data::Q, config::DSPConfig, τ::Quantity{T}, pars_filter::PropDict) where {Q <: Table, T<:Real}

DSP for ICPC detectors. It needs the decay time τ of the detector and the filter parameters pars_filter for the optimal filter parameters for the Trap, CUSP and ZAC filter.

Input data

The input data is a table with the following columns:

waveform: waveform data

baseline: baseline from FADC

timestamp: timestamp from FADC

eventnumber: event ID from FADC

daqenergy: energy from FADC

Output data

The output data is a table with the following columns:

blmean: baseline mean

blsigma: baseline sigma

blslope: baseline slope

bloffset: baseline offset

tailmean: tail mean after PZ correction

tailsigma: tail sigma after PZ correction

tailslope: tail slope after PZ correction

tailoffset: tail offset after PZ correction

t0: start time of waveform drift

t50: timepoint of 50% of waveform maximum

t80: timepoint of 80% of waveform maximum

t0_current: timepoint of current rise start

tail_τ: tail decay time

tail_mean: tail mean before PZ correction

tail_sigma: tail sigma before PZ correction

e_max: maximum of waveform

e_min: minimum of waveform

e_10410: energy of waveform with trapezoidal filter of 10µs rise time with 4µs flat-top

e_313: energy of waveform with trapezoidal filter of 3µs rise time with 1µs flat-top

e_10410_inv: maximum of inverted waveform with trapezoidal filter of 10µs rise time with 4µs flat-top

e_313_inv: maximum of inverted waveform with trapezoidal filter of 3µs rise time with 1µs flat-top

t0_inv: start time of inverted waveform drift

e_trap: energy of waveform with trapezoidal filter of optimized rise and flat-top time

e_cusp: energy of waveform with CUSP filter of optimized rise and flat-top time

e_zac: energy of waveform with ZAC filter of optimized rise and flat-top time

qdrift: Q-drift parameter

lq: LQ parameter

a: current maximum with optimal Savitzky-Golay filter length parameter

blfc: baseline from FADC

timestamp: timestamp from FADC

eventID_fadc: event ID from FADC

e_fc: energy from FADC

pretrace_diff: difference between first sample and baseline mean

rt1090: rise time between 10% and 90% of waveform maximum

rt1099: rise time between 10% and 99% of waveform maximum

rt9099: rise time between 90% and 99% of waveform maximum

drift_time: drift time between t0 and 90% of waveform maximum

inTrace_intersect: position of in-trace pile-up

inTrace_n: multiplicity of in-trace pile-up

n_sat_low: number of samples the waveform is saturated at low of FADC range

n_sat_high: number of samples the waveform is saturated at high of FADC range

n_sat_low_cons: number of consecutive samples the waveform is saturated at low of FADC range

n_sat_high_cons: number of consecutive samples the waveform is saturated at high of FADC range

LegendDSP.dsp_puls — Method

dsp_puls(data::Q, config::DSPConfig) where {Q <: Table}

DSP function for pulser processing.

Input data

The input data is a table with the following columns:

waveform: waveform data

baseline: baseline data

timestamp: timestamp data

eventnumber: event number data

daqenergy: energy data

Output data

The output data is a table with the following columns:

blmean: baseline mean

blsigma: baseline sigma

blslope: baseline slope

bloffset: baseline offset

t50: timepoint of 50% of waveform maximum

e_max: maximum of waveform

e_10410: energy of waveform with trapezoidal filter of 10µs rise time with 4µs flat-top

blfc: baseline from FADC

timestamp: timestamp

eventID_fadc: event number from FADC

e_fc: energy from FADC

LegendDSP.dsp_sg_optimization — Method

dsp_sg_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}, pars_filter::PropDict) where T<:Real

Optimize the Savitzky-Golay filter parameters for a given waveform set.

Returns

- `aoe`: Array of efficiency values for the given Savitzky-Golay filter parameters
- `e`: Array of energy values for the given Savitzky-Golay filter parameters
- `blmean`: Baseline mean value
- `blslope`: Baseline slope value

LegendDSP.dsp_sipm — Method

dsp_sipm(data::Q, config::PropDict, pars_threshold::PropDict)

DSP routine for SiPM data. It needs the threshold parameters from the threshold scan for the given SiPM channel as well as the discharge threshold parameters.

Input data

The input data is a table with the following columns:

waveform: waveform data

baseline: baseline data

timestamp: timestamp data

eventnumber: event number data

daqenergy: energy data

Output data

The output data is a table with the following columns:

blfc: baseline from FADC

timestamp: timestamp

eventID_fadc: event number from FADC

e_fc: energy from FADC

trig_pos: trigger positions from DSP

trig_max: trigger maxima from DSP

trig_pos_DC: trigger positions of discharges

trig_max_DC: trigger maxima of discharges

LegendDSP.dsp_trap_ft_optimization — Method

dsp_trap_ft_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}, rt::Quantity{T}) where T<:Real

Get energy grid values for given trap grid rise times while varying the flat-top time.

Returns

- `e_grid`: Array energy values for the given trap rise time grid at a given flat-top time grid.

LegendDSP.dsp_trap_rt_optimization — Method

dsp_trap_rt_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T},; ft::Quantity{T}=4.0u"µs") where T<:Real

Get ENC noise grid values for given trap grid rise times.

Returns

- `enc_trap_grid`: Array ENC noise values for the given trap rise time grid

LegendDSP.dsp_zac_ft_optimization — Method

dsp_zac_ft_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T}, rt::Quantity{T}) where T<:Real

Get energy grid values for given ZAC grid rise times while varying the flat-top time.

Returns: - e_grid: Array energy values for the given ZAC rise time grid at a given flat-top time grid.

LegendDSP.dsp_zac_rt_optimization — Method

dsp_zac_rt_optimization(wvfs::ArrayOfRDWaveforms, config::DSPConfig, τ::Quantity{T},; ft::Quantity{T}=4.0u"µs") where T<:Real

Get ENC noise grid values for given ZAC grid rise times.

Returns

- `enc_zac_grid`: Array ENC noise values for the given ZAC rise time grid

LegendDSP.extrema3points — Method

extrema3points(y1::T,y2::T,y3::T)::T where {T<:AbstractFloat}

Calculate the extrema of a parabola defined by three points.

LegendDSP.get_t0 — Method

get_t0(wvfs_pz::ArrayOfRDWaveforms, t0_threshold::T) where T<:Real

Get t0 for each waveform in wvfs_pz by using a fast asymetric trapezoidal filter and a fixed threshold at t0_threshold. The filter is truncated to the range 0µs to 60µs where the Ge trigger is expected in FlashCam.

LegendDSP.get_t50 — Method

get_t50(wvfs_pz::ArrayOfRDWaveforms, wvf_max::T) where T<:Real

Get t50 for each waveform in wvfs_pz by intersecting the waveform with a fixed threshold at wvf_max * 0.5.

LegendDSP.get_t80 — Method

get_t80(wvfs_pz::ArrayOfRDWaveforms, wvf_max::T) where T<:Real

Get t80 for each waveform in wvfs_pz by intersecting the waveform with a fixed threshold at wvf_max * 0.8.

LegendDSP.get_wvf_maximum — Function

get_wvf_maximum(signal::AbstractSamples, start::Real, stop::Real)
get_wvf_maximum(signal::RDWaveform, start::RealQuantity, stop::RealQuantity)

Get the maximum of a signal in the interval (start,stop) by quadaratic interpolation.

LegendDSP.saturation — Function

tailstats(signal::AbstractSamples, start::Real, stop::Real)
tailstats(signal::RDWaveform, start::RealQuantity, stop::RealQuantity)

Get statistics on the logarhithmic of the tail of a signal in the interval (start,stop).

LegendDSP.tailstats — Function

tailstats(signal::AbstractSamples, start::Real, stop::Real)
tailstats(signal::RDWaveform, start::RealQuantity, stop::RealQuantity)

Get statistics on the logarhithmic of the tail of a signal in the interval (start,stop) by fitting a linear function.